Use fixed width integers.
Patch by Raúl Marín Rodríguez.

Closes #3970
Closes https://github.com/postgis/postgis/pull/183


git-svn-id: http://svn.osgeo.org/postgis/trunk@16270 b70326c6-7e19-0410-871a-916f4a2858ee

+++ liblwgeom/cunit/cu_libgeom.c
@@ -257,7 +257,7 @@ static void test_lwgeom_from_gserialized(void)
 	GSERIALIZED *g;
 	char *in_ewkt;
 	char *out_ewkt;
-	int i = 0;
+	size_t i = 0;
 
 	char *ewkt[] =
 	{
@@ -698,7 +698,7 @@ static void test_f2d(void)
  */
 static void test_lwgeom_clone(void)
 {
-	int i;
+	size_t i;
 
 	char *ewkt[] =
 	{
